Revert "Kosmetyka"
This reverts commit d5daae8f09a958fdc923c3dbd86f4b37ae364d66.

+++ Colony.cpp
@@ -1,11 +1,10 @@
 #include "Colony.h"
 #include "Bunny.h"
-#include <time.h> // time()
 
 
 Colony::Colony()
 {
-    srand(time(NULL));
+    srand((int)time(NULL));
 }
 
 bool Colony::Cycle()
